# backend/docs/swagger.yaml
definitions:
...
  github_com_raids-lab_crater_internal_resputil.Response-any:
    properties:
      code:
        description: 依然保持 int (ErrorCode) 类型
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  github_com_raids-lab_crater_internal_resputil.Response-internal_handler_GpuAnalysisStatusResp:
    properties:
      code:
        description: 依然保持 int (ErrorCode) 类型
        type: integer
      data:
        $ref: '#/definitions/internal_handler.GpuAnalysisStatusResp'
      msg:
        type: string
    type: object
...
  github_com_raids-lab_crater_internal_resputil.Response-internal_handler_LLMConfigResp:
    properties:
      code:
        description: 依然保持 int (ErrorCode) 类型
        type: integer
      data:
        $ref: '#/definitions/internal_handler.LLMConfigResp'
      msg:
        type: string
    type: object
...
  github_com_raids-lab_crater_internal_resputil.Response-internal_handler_PrequeueConfigResp:
    properties:
      code:
        description: 依然保持 int (ErrorCode) 类型
        type: integer
      data:
        $ref: '#/definitions/internal_handler.PrequeueConfigResp'
      msg:
        type: string
    type: object
...
  github_com_raids-lab_crater_internal_resputil.Response-internal_handler_UserBanPolicyReq:
    properties:
      code:
        description: 依然保持 int (ErrorCode) 类型
        type: integer
      data:
        $ref: '#/definitions/internal_handler.UserBanPolicyReq'
      msg:
        type: string
    type: object
...
  github_com_raids-lab_crater_internal_resputil.Response-string:
    properties:
      code:
        description: 依然保持 int (ErrorCode) 类型
        type: integer
      data:
        type: string
      msg:
        type: string
    type: object
...
  internal_handler.GpuAnalysisStatusResp:
    properties:
      enabled:
        type: boolean
    type: object
...
  internal_handler.LLMConfigResp:
    properties:
      apiKey:
        type: string
      baseUrl:
        type: string
      modelName:
        type: string
    type: object
...
  internal_handler.PrequeueConfigResp:
    properties:
      activateTickerIntervalSeconds:
        type: integer
      backfillEnabled:
        type: boolean
      maxTotalActivationsPerRound:
        type: integer
      normalJobWaitingToleranceSeconds:
        type: integer
      prequeueCandidateSize:
        type: integer
      queueQuotaEnabled:
        type: boolean
    type: object
...
  internal_handler.SetGpuAnalysisStatusReq:
    properties:
      enable:
        type: boolean
    type: object
...
  internal_handler.UpdateLLMConfigReq:
    properties:
      apiKey:
        type: string
      baseUrl:
        type: string
      modelName:
        type: string
      validate:
        description: 是否立即校验连接
        type: boolean
    required:
    - baseUrl
    - modelName
    type: object
  internal_handler.UpdatePrequeueConfigReq:
    properties:
      activateTickerIntervalSeconds:
        type: integer
      backfillEnabled:
        type: boolean
      maxTotalActivationsPerRound:
        type: integer
      normalJobWaitingToleranceSeconds:
        type: integer
      prequeueCandidateSize:
        type: integer
      queueQuotaEnabled:
        type: boolean
    required:
    - activateTickerIntervalSeconds
    - backfillEnabled
    - maxTotalActivationsPerRound
    - normalJobWaitingToleranceSeconds
    - prequeueCandidateSize
    - queueQuotaEnabled
    type: object
...
  internal_handler.UserBanPolicyReq:
    properties:
      allowDatasetDownload:
        type: boolean
      allowImageBuild:
        type: boolean
      allowJobSubmission:
        type: boolean
      allowModelDownload:
        type: boolean
      allowPlatformAccess:
        type: boolean
    type: object
...
info:
  contact: {}
  description: This is the API server for Crater, a Multi-tenant AI Model Training
    Platform based on Kubernetes.
  title: Crater API
  version: 1.0.0
paths:
...
  /v1/admin/system-config/gpu-analysis:
    get:
      description: 查询当前系统是否开启了自动 GPU 资源滥用检测
      produces:
      - application/json
      responses:
        "200":
          description: 开关状态
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-internal_handler_GpuAnalysisStatusResp'
      security:
      - Bearer: []
      summary: 获取 GPU 分析功能开关状态
      tags:
      - SystemConfig
    put:
      consumes:
      - application/json
      description: 开启或关闭 GPU 检测。注意：尝试开启时，系统会强制检查 LLM 连接，如果连接不通，将无法开启。
      parameters:
      - description: 开关设置
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/internal_handler.SetGpuAnalysisStatusReq'
      produces:
      - application/json
      responses:
        "200":
          description: 设置成功
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-string'
        "400":
          description: LLM连接检查失败，无法开启
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: 设置 GPU 分析功能开关
      tags:
      - SystemConfig
  /v1/admin/system-config/llm:
    delete:
      description: 清空 LLM 配置（BaseURL, Key, Model）并强制关闭 GPU 分析功能
      produces:
      - application/json
      responses:
        "200":
          description: 重置成功
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-string'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: 重置 LLM 配置
      tags:
      - SystemConfig
    get:
      description: 获取当前系统配置的 LLM BaseURL, ModelName。出于安全考虑，API Key 可能会被脱敏显示。
      produces:
      - application/json
      responses:
        "200":
          description: 配置信息
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-internal_handler_LLMConfigResp'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: 获取 LLM 配置信息
      tags:
      - SystemConfig
    put:
      consumes:
      - application/json
      description: 更新 LLM 的连接信息。如果 validate 为 true，会尝试连接 /check 接口，失败则不保存。
      parameters:
      - description: 配置信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/internal_handler.UpdateLLMConfigReq'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-string'
        "400":
          description: 参数错误或校验失败
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: 更新 LLM 配置
      tags:
      - SystemConfig
  /v1/admin/system-config/prequeue:
    get:
      description: 获取当前回填提交开关、Crater 队内资源配额开关、普通作业等待忍耐时间和 watcher 运行参数
      produces:
      - application/json
      responses:
        "200":
          description: 配置
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-internal_handler_PrequeueConfigResp'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: 获取新版排队配置
      tags:
      - SystemConfig
    put:
      consumes:
      - application/json
      description: 更新回填提交开关、Crater 队内资源配额开关、普通作业等待忍耐时间和 watcher 运行参数
      parameters:
      - description: 配置
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/internal_handler.UpdatePrequeueConfigReq'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-string'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: 更新新版排队配置
      tags:
      - SystemConfig
  /v1/admin/system-config/user-ban:
    get:
      description: 获取封禁用户仍被允许使用的平台能力
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-internal_handler_UserBanPolicyReq'
      security:
      - Bearer: []
      summary: 获取用户封禁能力策略
      tags:
      - SystemConfig
    put:
      consumes:
      - application/json
      description: 设置封禁用户是否仍可访问平台、提交作业、制作镜像以及下载模型或数据集
      parameters:
      - description: 封禁能力策略
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/internal_handler.UserBanPolicyReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-internal_handler_UserBanPolicyReq'
      security:
      - Bearer: []
      summary: 更新用户封禁能力策略
      tags:
      - SystemConfig
...
securityDefinitions:
  Bearer:
    description: 访问 /login 并获取 TOKEN 后，填入 'Bearer ${TOKEN}' 以访问受保护的接口
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"